revert the move of DirectMemoryUtils and make calculateClientMemoryLimit method back to ThreadRuntimeFactory
