+++ pulsar-functions/runtime/src/main/java/org/apache/pulsar/functions/runtime/thread/ThreadRuntimeFactory.java
@@ -18,7 +18,7 @@
  */
 package org.apache.pulsar.functions.runtime.thread;
 
-import static org.apache.pulsar.common.functions.MemoryLimit.calculateClientMemoryLimit;
+import static com.google.common.base.Preconditions.checkArgument;
 import java.util.Optional;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -28,6 +28,7 @@ import org.apache.pulsar.client.api.ClientBuilder;
 import org.apache.pulsar.client.api.PulsarClient;
 import org.apache.pulsar.client.api.PulsarClientException;
 import org.apache.pulsar.common.functions.MemoryLimit;
+import org.apache.pulsar.common.util.DirectMemoryUtils;
 import org.apache.pulsar.common.util.Reflections;
 import org.apache.pulsar.functions.auth.FunctionAuthProvider;
 import org.apache.pulsar.functions.instance.AuthenticationConfig;
@@ -140,6 +141,37 @@ public class ThreadRuntimeFactory implements RuntimeFactory {
         this.functionsManager = functionsManager;
     }
 
+    private Optional<Long> calculateClientMemoryLimit(Optional<MemoryLimit> memoryLimit) {
+        if (memoryLimit.isPresent()) {
+
+            Long absolute = memoryLimit.get().getAbsoluteValue();
+            Double percentOfDirectMem = memoryLimit.get().getPercentOfMaxDirectMemory();
+            if (absolute != null) {
+                checkArgument(absolute > 0, "Absolute memory limit for Pulsar client has to be positive");
+            }
+            if (percentOfDirectMem != null) {
+                checkArgument(percentOfDirectMem > 0 && percentOfDirectMem <= 100,
+                        "Percent of max direct memory limit for Pulsar client must be between 0 and 100");
+            }
+
+            if (absolute != null && percentOfDirectMem != null) {
+                return Optional.of(Math.min(absolute, getBytesPercentDirectMem(percentOfDirectMem)));
+            }
+
+            if (absolute != null) {
+                return Optional.of(absolute);
+            }
+
+            if (percentOfDirectMem != null) {
+                return Optional.of(getBytesPercentDirectMem(percentOfDirectMem));
+            }
+        }
+        return Optional.empty();
+    }
+
+    private long getBytesPercentDirectMem(double percent) {
+        return (long) (DirectMemoryUtils.jvmMaxDirectMemory() * (percent / 100));
+    }
 
 
     @Override
